+++ serverless/src/lambda/redirect.ts
