+++ serverless/src/lambda/tracing.ts
@@ -1,6 +1,6 @@
 import { LambdaFunction } from "./layer";
 import { Configuration } from "./env";
-import { Resources } from "./index";
+import { Resources } from "../types";
 import log from "loglevel";
 
 const FN_GET_ATT = "Fn::GetAtt";
